+++ awc/src/lib.rs
@@ -38,6 +38,7 @@ pub mod test;
 pub mod ws;
 
 pub use self::builder::ClientBuilder;
+pub use self::connect::BoxedSocket;
 pub use self::request::ClientRequest;
 pub use self::response::{ClientResponse, JsonBody, MessageBody};