+++ src/middleware/mod.rs
@@ -11,6 +11,3 @@ mod normalize;
 pub use self::defaultheaders::DefaultHeaders;
 pub use self::logger::Logger;
 pub use self::normalize::NormalizePath;
-
-#[cfg(feature = "secure-cookies")]
-pub mod identity;